fix(types): 9 mypy errors in training + fusion + edge
F-A partial fix (3 files cleared, -9 errors):

**edge/aer_router.py (3 → 0):**
- `self._process = None` → `self._process: subprocess.Popen[bytes] | None = None` so the Popen assignment in start() doesn't conflict with mypy's inferred `None` type.
- `start(self, build: bool = True) -> None:` and `stop(self) -> None:`.

**fusion/sensor_fusion.py (3 → 0):**
- `CrossModalAttention._scc`: same `float(np.mean(...))` coerce pattern as sc_runtime / _compute_scc. numpy's Generator mean returns a `_FloatOp` union mypy can't feed into `min(pa, pb)` / `float()`; coercing at assignment time resolves the chain. `max(0, ...)` → `max(0.0, ...)`.

**training/snn_modules.py (3 → 0):**
- ModuleList cast pattern (same as qat/torch_qat): `cast(torch.nn.Linear, lin)` / `cast(LIFCell, self.lifs[i])` at each access site. `to_sc_weights` extracts `lin_typed = cast(torch.nn.Linear, lin)` once per iteration since the element is accessed three times.

Local mypy: 84 → 75 total (-9). 17 files now fully green. Next targets
(smaller files, ≤3 errors each):
  - _native/learning_bridge.py (5 — deferred; complex autograd.Function apply signature)
  - experimental/alternative_path.py (3)
  - debug/sc_scope.py (3 — deferred; depends on Codex T0.2 stochastic_doctor)
  - bridges/{quantum_annealing, dna_mapper}.py (3+3)
  - analog_bridge.py (3), accel/jax_backend.py (3)
  - accel/julia/solvers/__init__.py (3)
  - solvers/{symplectic, stiff}.py (2+2)
  - pipeline/training.py (2), _native/core_engine_bridge.py (2)
  - formal/lean_bridge.py (2)

Stays strictly outside Codex's active T0.1 + T0.2 scope.

Session log: .coordination/sessions/claude_2026-04-20T1902_ci_full_fix_4_failures.md

+++ src/sc_neurocore/fusion/sensor_fusion.py
@@ -124,16 +124,16 @@ class BitstreamDecorrelator:
         """Compute stochastic cross-correlation between two bitstreams."""
         a_flat = a.flatten().astype(np.float64)
         b_flat = b.flatten().astype(np.float64)
-        pa = np.mean(a_flat)
-        pb = np.mean(b_flat)
-        p_and = np.mean(a_flat * b_flat)
+        pa = float(np.mean(a_flat))
+        pb = float(np.mean(b_flat))
+        p_and = float(np.mean(a_flat * b_flat))
         num = p_and - (pa * pb)
         if abs(num) < 1e-12:
             return 0.0
-        denom = (min(pa, pb) - pa * pb) if num > 0 else (pa * pb - max(0, pa + pb - 1))
+        denom = (min(pa, pb) - pa * pb) if num > 0 else (pa * pb - max(0.0, pa + pb - 1.0))
         if abs(denom) < 1e-12:
             return 0.0
-        return float(max(-1.0, min(1.0, num / denom)))
+        return max(-1.0, min(1.0, num / denom))
 
 
 class CrossModalAttention:
